+++ calibration-app/app/src/main/java/org/dpppt/android/calibration/controls/ControlsFragment.java
@@ -307,7 +307,7 @@ public class ControlsFragment extends Fragment {
 		Button buttonReportFake = view.findViewById(R.id.home_button_report_fake);
 		buttonReportFake.setOnClickListener(
 				v -> {
-					DP3T.sendFakeInfectedRequest(getContext(), null);
+					DP3T.sendFakeInfectedRequest(getContext(), null, null, null);
 				});
 	}