+++ calibration-app/app/src/main/java/org/dpppt/android/calibration/handshakes/BackendUserBucketRepository.java
@@ -10,9 +10,6 @@
 package org.dpppt.android.calibration.handshakes;
 
 import android.content.Context;
-import androidx.annotation.NonNull;
-
-import java.io.IOException;
 
 import org.dpppt.android.calibration.MainApplication;
 import org.dpppt.android.sdk.backend.SignatureException;
@@ -20,6 +17,9 @@ import org.dpppt.android.sdk.internal.backend.Repository;
 import org.dpppt.android.sdk.internal.backend.ServerTimeOffsetException;
 import org.dpppt.android.sdk.internal.backend.StatusCodeException;
 
+import java.io.IOException;
+
+import androidx.annotation.NonNull;
 import okhttp3.ResponseBody;
 import retrofit2.Response;
 import retrofit2.Retrofit;
@@ -47,7 +47,7 @@ public class BackendUserBucketRepository implements Repository {
 		if (response.isSuccessful() && response.body() != null) {
 			return response.body();
 		} else {
-			throw new StatusCodeException(response.raw());
+			throw new StatusCodeException(response.raw(), response.errorBody());
 		}
 	}