+++ dp3t-sdk/sdk/src/androidTest/java/org/dpppt/android/sdk/internal/nearby/TestGoogleExposureClient.java
@@ -22,22 +22,27 @@ import java.util.List;
 
 import com.google.android.gms.common.api.Api;
 import com.google.android.gms.common.api.internal.ApiKey;
-import com.google.android.gms.nearby.exposurenotification.ExposureConfiguration;
-import com.google.android.gms.nearby.exposurenotification.ExposureInformation;
-import com.google.android.gms.nearby.exposurenotification.ExposureNotificationClient;
-import com.google.android.gms.nearby.exposurenotification.ExposureSummary;
-import com.google.android.gms.nearby.exposurenotification.TemporaryExposureKey;
+import com.google.android.gms.nearby.exposurenotification.*;
 import com.google.android.gms.tasks.Task;
 
 import org.dpppt.android.sdk.internal.util.Json;
+import org.dpppt.android.sdk.models.DayDate;
+import org.dpppt.android.sdk.util.DateUtil;
 
 public class TestGoogleExposureClient implements ExposureNotificationClient {
 
 	private Context context;
 	private int provideDiagnosisKeysCounter = 0;
+	private boolean currentDayKeyReleased = false;
+	private long time = System.currentTimeMillis();
 
 	public TestGoogleExposureClient(Context context) {
 		this.context = context;
+	}
+
+	public TestGoogleExposureClient(Context context, boolean currentDayKeyReleased) {
+		this.context = context;
+		this.currentDayKeyReleased = currentDayKeyReleased;
 	}
 
 	@Override
@@ -57,7 +62,18 @@ public class TestGoogleExposureClient implements ExposureNotificationClient {
 
 	@Override
 	public Task<List<TemporaryExposureKey>> getTemporaryExposureKeyHistory() {
-		return new DummyTask<>(new ArrayList<>());
+		ArrayList<TemporaryExposureKey> temporaryExposureKeys = new ArrayList<>();
+		for (int i = 1; i<14; i++){
+			temporaryExposureKeys.add(new TemporaryExposureKey.TemporaryExposureKeyBuilder()
+					.setRollingStartIntervalNumber(DateUtil.getRollingStartNumberForDate(new DayDate(time).subtractDays(i)))
+					.build());
+		}
+		if (currentDayKeyReleased) {
+			temporaryExposureKeys.add(new TemporaryExposureKey.TemporaryExposureKeyBuilder()
+					.setRollingStartIntervalNumber(DateUtil.getRollingStartNumberForDate(new DayDate(time)))
+					.build());
+		}
+		return new DummyTask<>(temporaryExposureKeys);
 	}
 
 	@Override
@@ -86,6 +102,11 @@ public class TestGoogleExposureClient implements ExposureNotificationClient {
 	}
 
 	@Override
+	public Task<List<ExposureWindow>> getExposureWindows(String s) {
+		return new DummyTask<>(new ArrayList<>());
+	}
+
+	@Override
 	public Task<ExposureSummary> getExposureSummary(String s) {
 		return new DummyTask<>(new ExposureSummary.ExposureSummaryBuilder().build());
 	}
@@ -102,6 +123,10 @@ public class TestGoogleExposureClient implements ExposureNotificationClient {
 
 	public int getProvideDiagnosisKeysCounter() {
 		return provideDiagnosisKeysCounter;
+	}
+
+	public void setTime(long time){
+		this.time = time;
 	}
 
 	public static class ExposureTestParameters {